Added support for nodejs
Node JS will require lodash to be explicitly loaded so just invoking _ will not be suffice.  We must allow the user to provide us with a lodash.

Also the name _ can do funny things inside of Node JS when in interactive mode.  I've replaced it with more explicit "lodash" for all but the absolutely necessary cases.