Added mode / modes
Added mode / modes.  It is better than simple-statistics in that it is mathematically correct.  There's will return the lowest most repeated value.  This one returns a set of the most frequent numbers.  It also allows for customized clustering so that we can find the mode of numbers that are unique in insignificant ways.

See this for information regarding mode being a set rather than a single value.  http://mathforum.org/library/drmath/view/61375.html